Refactoring components and creating Netlify CMS initial configs
