+++ content/aplicacoes-serveless-na-aws-com-node.md
@@ -1,0 +1,10 @@
+---
+thumbnail: assets/img/courses/aws.png
+title: Aplicações Serveless na AWS com Node.Js
+description: Seja um profissional completo em tecnologias cloud computing
+level: Avançado
+authorImage: assets/img/authors/erick-wendel.png
+author: Erick Wendel
+price: 400
+discount: 329,90
+---
