+++ package.json
@@ -9,7 +9,9 @@
     "babel-plugin-styled-components": "^1.10.7",
     "gatsby": "^2.20.12",
     "gatsby-image": "^2.3.1",
+    "gatsby-plugin-google-analytics": "^2.3.0",
     "gatsby-plugin-manifest": "^2.3.3",
+    "gatsby-plugin-netlify-cms": "^4.3.0",
     "gatsby-plugin-offline": "^3.1.2",
     "gatsby-plugin-prefetch-google-fonts": "^1.4.3",
     "gatsby-plugin-react-helmet": "^3.2.1",
@@ -17,6 +19,7 @@
     "gatsby-plugin-styled-components": "^3.2.1",
     "gatsby-source-filesystem": "^2.2.2",
     "gatsby-transformer-sharp": "^2.4.3",
+    "netlify-cms-app": "^2.12.11",
     "prop-types": "^15.7.2",
     "react": "^16.12.0",
     "react-dom": "^16.12.0",
