+++ src/templates/info-cards.js
@@ -1,0 +1,21 @@
+import startup from "../images/icons/startup.svg"
+import terminal from "../images/icons/code-terminal.svg"
+import brain from "../images/icons/brain.svg"
+
+export const infoCards = [
+  {
+    icon: startup,
+    title: "Aprenda na prática",
+    description: "Aprenda tão rápido como um foguete!",
+  },
+  {
+    icon: terminal,
+    title: "Code Review",
+    description: "Cursos com direito a code review",
+  },
+  {
+    icon: brain,
+    title: "Divirta-se",
+    description: "Aprenda com muita diversão !",
+  },
+]
