+++ static/admin/config.yml
@@ -1,0 +1,24 @@
+backend:
+  name: github
+  repo: lancaluis/cursos-landing-v1
+  branch: v2
+  site_domain: sleepy-carson-9cbe50.netlify.app
+
+publish_mode: editorial_workflow
+media_folder: "static/assets/img"
+public_folder: "assets/img"
+
+collections:
+  - name: "courses" # Used in routes, e.g., /admin/collections/blog
+    label: "Courses" # Used in the UI
+    folder: "content" # The path to the folder where the documents are stored
+    create: true # Allow users to create new documents in this collection
+    fields: # The fields for each document, usually in front matter
+      - { label: "Thumbnail", name: "thumbnail", widget: "image" }
+      - { label: "Title", name: "title", widget: "string" } 
+      - { label: "Description", name: "description", widget: "string" }
+      - { label: "Level", name: "level", widget: "select", options: ["Iniciante", "Intermediário", "Avançado"] }
+      - { label: "Author Image", name: "authorImage", widget: "image" }
+      - { label: "Author", name: "author", widget: "string" }
+      - { label: "Price", name: "price", widget: "string" }
+      - { label: "Discount", name: "discount", widget: "string" }
