+++ static/robots.txt
@@ -1,0 +1,1 @@
+User-agent: *
